+++ web/usuaris/src/__mocks__/handlers.ts
@@ -6,11 +6,19 @@ export const handlers = [
   http.get(urls.getUsers, ({ request }) => {
     const url = new URL(request.url)
     const itinerarySlug = url.searchParams.get('itinerarySlug')
+    const startDate = url.searchParams.get('startDate')
+    const endDate = url.searchParams.get('endDate')
     const name = url.searchParams.get('name')
     const dni = url.searchParams.get('dni')
 
     if (
       itinerarySlug === 'frontend-react' &&
+      ((startDate &&
+        startDate <= '2023/11/06 00:00:00.000' &&
+        endDate &&
+        endDate >= '2023/11/06 00:00:00.000') ||
+        (startDate && startDate <= '2023/11/06 00:00:00.000') ||
+        (endDate && endDate >= '2023/11/06 00:00:00.000')) &&
       (name?.includes('marc') || dni?.includes('marc'))
     ) {
       return HttpResponse.json(
@@ -20,7 +28,7 @@ export const handlers = [
             name: 'Marc Bofill',
             dni: '87654321B',
             status: 'ACTIVE',
-            createdAt: '2023/11/05 00:00:00.000',
+            createdAt: '2023/11/06 00:00:00.000',
             itineraryName: 'Frontend React',
           },
         ],
@@ -43,8 +51,12 @@ export const handlers = [
           name: 'Marc Bofill',
           dni: '87654321B',
           status: 'ACTIVE',
+<<<<<<< HEAD
           role: UserRole.REGISTERED,
           createdAt: '2023/11/05 00:00:00.000',
+=======
+          createdAt: '2023/11/06 00:00:00.000',
+>>>>>>> da50621207ec71e4a060584627ccd3e7347b8f0d
           itineraryName: 'Frontend React',
         },
         {
@@ -52,8 +64,12 @@ export const handlers = [
           name: 'Montserrat Capdevila',
           dni: '45678912C',
           status: 'BLOCKED',
+<<<<<<< HEAD
           role: UserRole.REGISTERED,
           createdAt: '2023/11/05 00:00:00.000',
+=======
+          createdAt: '2023/11/07 00:00:00.000',
+>>>>>>> da50621207ec71e4a060584627ccd3e7347b8f0d
           itineraryName: 'Fullstack Php',
         },
         {
@@ -61,8 +77,12 @@ export const handlers = [
           name: 'Anna Brull',
           dni: '45678912D',
           status: 'BLOCKED',
+<<<<<<< HEAD
           role: UserRole.ADMIN,
           createdAt: '2023/11/05 00:00:00.000',
+=======
+          createdAt: '2023/11/08 00:00:00.000',
+>>>>>>> da50621207ec71e4a060584627ccd3e7347b8f0d
           itineraryName: 'Frontend React',
         },
         {
@@ -70,8 +90,12 @@ export const handlers = [
           name: 'Marc Serra',
           dni: '12378912D',
           status: 'BLOCKED',
+<<<<<<< HEAD
           role: UserRole.ADMIN,
           createdAt: '2023/11/05 00:00:00.000',
+=======
+          createdAt: '2023/11/09 00:00:00.000',
+>>>>>>> da50621207ec71e4a060584627ccd3e7347b8f0d
           itineraryName: 'Frontend Angular',
         },
       ],