+++ web/usuaris/src/components/organisms/FiltersWidget.tsx
@@ -14,52 +14,43 @@ const FiltersContainer = styled(FlexBox)`
 
 type TFiltersWidget = {
   filters: TFilters
-  handleFilters: (filters: TFilters) => void
+  setFilters: (filters: TFilters) => void
 }
 
-export const FiltersWidget: FC<TFiltersWidget> = ({
-  filters,
-  handleFilters,
-}) => {
+export const FiltersWidget: FC<TFiltersWidget> = ({ filters, setFilters }) => {
   const { t } = useTranslation()
 
   const handleItinerary = (itinerary: TItinerary | undefined) => {
-    if (itinerary !== undefined) {
-      handleFilters({ ...filters, itinerarySlug: itinerary.slug })
-    } else {
-      const newFilters = { ...filters }
-      delete newFilters.itinerarySlug
-      handleFilters(newFilters)
-    }
+    setFilters({ ...filters, itinerarySlug: itinerary?.slug })
   }
   const handleRole = (role: TRol | undefined) => {
     if (role === undefined) {
       const newFilters = { ...filters }
       delete newFilters.role
-      handleFilters(newFilters)
+      setFilters(newFilters)
     } else {
-      handleFilters({ ...filters, role: role.slug })
+      setFilters({ ...filters, role: role.slug })
     }
   }
 
   const handleSearch = useCallback(
     (searchValue: string) => {
-      if (searchValue !== '') {
-        handleFilters({
-          ...filters,
-          name: searchValue,
-          dni: searchValue,
-        })
-      } else {
-        const newFilters = { ...filters }
-        delete newFilters.name
-        delete newFilters.dni
-        handleFilters(newFilters)
-      }
+      setFilters({
+        ...filters,
+        name: searchValue,
+        dni: searchValue,
+      })
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
+    [filters]
   )
+  const handleDates = (startDate: Date | null, endDate: Date | null) => {
+    setFilters({
+      ...filters,
+      startDate: startDate?.toISOString(),
+      endDate: endDate?.toISOString(),
+    })
+  }
 
   return (
     <FiltersContainer
@@ -76,6 +67,7 @@ export const FiltersWidget: FC<TFiltersWidget> = ({
         placeholderEndDate={t('Hasta...')}
         dateFormat="dd/MM/yyyy"
         calendarLanguage={t('calendarLanguage')}
+        handleDates={handleDates}
       />
       <Search
         id="usersSearch"