+++ web/usuaris/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.13.0] - 2024-05-02
+## [0.14.0] - 2024-05-02
 
 ### Added
 
@@ -10,7 +10,16 @@ All notable changes to this project will be documented in this file.
 - Added Roles filter feature
 - Add Test for Roles
   
-  
+
+## [0.13.0] - 2024-04-29
+
+### Added
+
+- Added close icon to deselect dates in DateRange
+- Added dates to filters to fetch users in FiltersWidget
+- Changed qs to handle null, undefined and empty string
+
+
 ## [0.12.0] - 2024-04-29
 
 ### Added